all_requirements.txt
boto>=2.36.0
boto3>=1.6.16
botocore>=1.12.13
cryptography>=2.3.0
docker>=2.5.1
Jinja2>=2.7.3
mock
pytz
future
wrapt
pyaml
werkzeug
airspeed==0.5.10
amazon-kclpy==1.5.0
awscli>=1.14.18
coverage>=4.0.3
docopt>=0.6.2
elasticsearch==6.2.0
nose>=1.3.7
psutil==5.4.8
pyyaml>=3.13
flask==0.12.4
six>=1.12.0
jsii~=0.14.0
aws-cdk.assets==1.0.0
aws-cdk.assets-docker==0.35.0
aws-cdk.aws-apigateway==1.0.0
aws-cdk.aws-applicationautoscaling==1.0.0
aws-cdk.aws-autoscaling==1.0.0
aws-cdk.aws-autoscaling-common==1.0.0
aws-cdk.aws-autoscaling-hooktargets==1.0.0
aws-cdk.aws-certificatemanager==1.0.0
aws-cdk.aws-cloudformation==1.0.0
aws-cdk.aws-cloudfront==1.0.0
aws-cdk.aws-cloudwatch==1.0.0
aws-cdk.aws-codebuild==1.0.0
aws-cdk.aws-codecommit==1.0.0
aws-cdk.aws-codepipeline==1.0.0
aws-cdk.aws-dynamodb==1.0.0
aws-cdk.aws-ec2==1.0.0
aws-cdk.aws-ecr==1.0.0
aws-cdk.aws-ecr-assets==1.0.0
aws-cdk.aws-ecs==1.0.0
aws-cdk.aws-ecs-patterns==1.0.0
aws-cdk.aws-efs==1.0.0
aws-cdk.aws-elasticloadbalancing==1.0.0
aws-cdk.aws-elasticloadbalancingv2==1.0.0
aws-cdk.aws-events==1.0.0
aws-cdk.aws-events-targets==1.0.0
aws-cdk.aws-iam==1.0.0
aws-cdk.aws-kinesis==1.0.0
aws-cdk.aws-kms==1.0.0
aws-cdk.aws-lambda==1.0.0
aws-cdk.aws-lambda-event-sources==1.0.0
aws-cdk.aws-logs==1.0.0
aws-cdk.aws-route53==1.0.0
aws-cdk.aws-route53-targets==1.0.0
aws-cdk.aws-s3==1.0.0
aws-cdk.aws-s3-assets==1.0.0
aws-cdk.aws-s3-notifications==1.0.0
aws-cdk.aws-sagemaker==1.0.0
aws-cdk.aws-secretsmanager==1.0.0
aws-cdk.aws-servicediscovery==1.0.0
aws-cdk.aws-sns==1.0.0
aws-cdk.aws-sns-subscriptions==1.0.0
aws-cdk.aws-sqs==1.0.0
aws-cdk.aws-ssm==1.0.0
aws-cdk.aws-stepfunctions==1.0.0
aws-cdk.aws-stepfunctions-tasks==1.0.0
aws-cdk.cdk==0.36.1
aws-cdk.core==1.0.0
aws-cdk.custom-resources==1.0.0
aws-cdk.cx-api==1.0.0
aws-cdk.region-info==1.0.0
